+++ src/senaite/patient/setuphandlers.py
@@ -18,20 +18,17 @@
 # Copyright 2020 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-import time
-
-import transaction
 from bika.lims import api
-from bika.lims.catalog import CATALOG_ANALYSIS_REQUEST_LISTING
-from bika.lims.catalog.catalog_utilities import addZCTextIndex
 from plone.registry.interfaces import IRegistry
 from Products.DCWorkflow.Guard import Guard
-from Products.ZCatalog.ProgressHandler import ZLogHandler
+from senaite.core.catalog import SAMPLE_CATALOG
+from senaite.core.setuphandlers import setup_core_catalogs
+from senaite.core.setuphandlers import setup_other_catalogs
 from senaite.core.workflow import SAMPLE_WORKFLOW
 from senaite.patient import PRODUCT_NAME
 from senaite.patient import logger
 from senaite.patient import permissions
-from senaite.patient.config import PATIENT_CATALOG
+from senaite.patient.catalog.patient_catalog import PatientCatalog
 from zope.component import getUtility
 
 PROFILE_ID = "profile-{}:default".format(PRODUCT_NAME)
@@ -40,26 +37,21 @@ PROFILE_ID = "profile-{}:default".format(PRODUCT_NAME)
 # Default: 300 (5 minutes)
 MAX_SEC_THRESHOLD = 300
 
-# Tuples of (type, [catalog])
-CATALOGS_BY_TYPE = [
-]
+CATALOGS = (
+    PatientCatalog,
+)
 
-# Tuples of (catalog, index_name, index_type)
+# Tuples of (catalog, index_name, index_attribute, index_type)
 INDEXES = [
-    (CATALOG_ANALYSIS_REQUEST_LISTING, "is_temporary_mrn", "BooleanIndex"),
-    (CATALOG_ANALYSIS_REQUEST_LISTING, "medical_record_number", "KeywordIndex"),
-    (PATIENT_CATALOG, "patient_mrn", "FieldIndex"),
-    (PATIENT_CATALOG, "patient_email", "FieldIndex"),
-    (PATIENT_CATALOG, "patient_fullname", "FieldIndex"),
-    (PATIENT_CATALOG, "patient_searchable_text", "TextIndexNG3"),
+    (SAMPLE_CATALOG, "is_temporary_mrn", "", "BooleanIndex"),
+    (SAMPLE_CATALOG, "medical_record_number", "", "KeywordIndex"),
 ]
 
 # Tuples of (catalog, column_name)
 COLUMNS = [
-    (CATALOG_ANALYSIS_REQUEST_LISTING, "isMedicalRecordTemporary"),
-    (CATALOG_ANALYSIS_REQUEST_LISTING, "getMedicalRecordNumberValue"),
-    (CATALOG_ANALYSIS_REQUEST_LISTING, "getPatientFullName"),
-    (PATIENT_CATALOG, "mrn"),
+    (SAMPLE_CATALOG, "isMedicalRecordTemporary"),
+    (SAMPLE_CATALOG, "getMedicalRecordNumberValue"),
+    (SAMPLE_CATALOG, "getPatientFullName"),
 ]
 
 NAVTYPES = [
@@ -225,6 +217,13 @@ def post_uninstall(portal_setup):
     logger.info("{} uninstall handler [DONE]".format(PRODUCT_NAME.upper()))
 
 
+def setup_catalogs(portal):
+    """Setup patient catalogs
+    """
+    setup_core_catalogs(portal, catalog_classes=CATALOGS)
+    setup_other_catalogs(portal, indexes=INDEXES, columns=COLUMNS)
+
+
 def add_patient_folder(portal):
     """Adds the initial Patient folder
     """
@@ -273,93 +272,6 @@ def setup_id_formatting(portal, format_definition=None):
         ids.append(record)
     ids.append(format_definition)
     bs.setIDFormatting(ids)
-
-
-def setup_catalogs(portal):
-    """Setup catalogs
-    """
-    logger.info("Setup Catalogs ...")
-
-    # Setup catalogs by type
-    logger.info("Setup Catalogs by type ...")
-    for type_name, catalogs in CATALOGS_BY_TYPE:
-        at = api.get_tool("archetype_tool")
-        # get the current registered catalogs
-        current_catalogs = at.getCatalogsByType(type_name)
-        # get the desired catalogs this type should be in
-        desired_catalogs = map(api.get_tool, catalogs)
-        # check if the catalogs changed for this portal_type
-        if set(desired_catalogs).difference(current_catalogs):
-            # fetch the brains to reindex
-            brains = api.search({"portal_type": type_name})
-            # updated the catalogs
-            at.setCatalogsByType(type_name, catalogs)
-            logger.info("Assign '%s' type to Catalogs %s" %
-                        (type_name, catalogs))
-            for brain in brains:
-                obj = api.get_object(brain)
-                logger.info("Reindexing '%s'" % repr(obj))
-                obj.reindexObject()
-
-    # Setup catalog indexes
-    logger.info("Setup indexes ...")
-    to_index = {}
-    for catalog, name, meta_type in INDEXES:
-        c = api.get_tool(catalog)
-        indexes = c.indexes()
-        if name in indexes:
-            logger.info("Index '%s' already in Catalog [SKIP]" % name)
-            continue
-
-        logger.info("Adding Index '%s' for field '%s' to catalog '%s"
-                    % (meta_type, name, catalog))
-        if meta_type == "ZCTextIndex":
-            addZCTextIndex(c, name)
-        else:
-            c.addIndex(name, meta_type)
-
-        if catalog not in to_index:
-            to_index[catalog] = [name, ]
-        else:
-            to_index[catalog].append(name)
-
-        logger.info("Added Index '%s' for field '%s' to catalog [DONE]"
-                    % (meta_type, name))
-
-    for catalog, names in to_index.items():
-        start = time.time()
-        logger.info("Indexing new index/es from {}: {} ..."
-                    .format(catalog, ", ".join(names)))
-        handler = ZLogHandler(steps=100)
-        c = api.get_tool(catalog)
-        c.reindexIndex(names, None, handler)
-        end = time.time()
-        if (end-start) > MAX_SEC_THRESHOLD:
-            commit_transaction(portal)
-
-    # Setup catalog metadata columns
-    for catalog, name in COLUMNS:
-        c = api.get_tool(catalog)
-        if name not in c.schema():
-            logger.info("Adding Column '%s' to catalog '%s' ..."
-                        % (name, catalog))
-            c.addColumn(name)
-            logger.info("Added Column '%s' to catalog '%s' [DONE]"
-                        % (name, catalog))
-        else:
-            logger.info("Column '%s' already in catalog '%s'  [SKIP]"
-                        % (name, catalog))
-            continue
-    logger.info("Setup Catalogs [DONE]")
-
-
-def commit_transaction(portal):
-    start = time.time()
-    logger.info("Commit transaction ...")
-    transaction.commit()
-    end = time.time()
-    logger.info("Commit transaction ... Took {:.2f}s [DONE]"
-                .format(end - start))
 
 
 def setup_workflow(portal):