+++ docs/changelog.rst
@@ -4,6 +4,6 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #14 Compatibility with Senaite catalog migration
 - #8 Added patient workflow and managed permissions
 - #7 Added upgrade step handler for senaite.patient
-- Initial Release